Move aiohttp_sse_client to common to avoid unnecessary imports
