+++ model-engine/model_engine_server/domain/use_cases/llm_model_endpoint_use_cases.py
@@ -218,6 +218,8 @@ _SUPPORTED_MODELS_BY_FRAMEWORK = {
             "llama-3-1-70b-instruct",
             "llama-3-1-405b",
             "llama-3-1-405b-instruct",
+            "llama-3-2-11b-vision-instruct",
+            "llama-3-2-90b-vision-instruct",
             "falcon-7b",
             "falcon-7b-instruct",
             "falcon-40b",
@@ -895,7 +897,10 @@ class CreateLLMModelBundleV1UseCase:
                         healthcheck_route="/health",
                         predict_route="/predict",
                         streaming_predict_route="/stream",
-                        extra_routes=[OPENAI_CHAT_COMPLETION_PATH, OPENAI_COMPLETION_PATH],
+                        extra_routes=[
+                            OPENAI_CHAT_COMPLETION_PATH,
+                            OPENAI_COMPLETION_PATH,
+                        ],
                         env={},
                     ),
                     metadata={},
